requirements.txt
discord.py
Flask
PyYAML
Werkzeug
bcrypt
python-dotenv
Flask-Session
cryptography
requests
Pillow
pytesseract
beautifulsoup4
jinja2
itsdangerous
click
numpy
opencv-python
